+++ dfmcontrol_C/Tests/Helper/TestErrorRate.h
@@ -1,1 +1,1 @@
-float ErrorRate(int bitsize, int genes, int individuals);
+double ErrorRate(int bitsize, int genes, int individuals);
